Change R4640 so it's a subclass of R4650 rather than MIPS3
The R4640 is essentially the same product as the R4650 but with a 32-bit bus width. They share the same PRId. This change makes it easier to pull upstream changes without having to copy code from R4650 into R4640.

This should work as long as the R4650 implementation stays 32-bit. It's 32-bit now but there's a note saying "Should be 64 bits". We'll deal with that if this changes.