Handle misaligned code fetch
Windows CE seems to rely a lot on processor exceptions to make kernel API calls. TLB mismatches is one of the ways it calls the kernel but it also jumps to addresses not aligned to 32-bit boundries so this code handles those cases. Limiting this to the RM5230 processor flavor so it doesn't disturb other drivers.